rosecore/requirements.txt
pipenv
gunicorn
django
psycopg2-binary==2.8.5
todoist-python==8.1.3
django-extensions
python-dateutil
pytz
tblib
